+++ spec/bourbon/library/font_stacks_spec.rb
@@ -21,9 +21,11 @@ describe "font-stacks" do
                     '"Lucida Typewriter", monospace'
       monaco = '"monaco", "Consolas", "Lucida Console", monospace'
 
-      system = '-apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", ' +
-               '"Oxygen", "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", ' +
-               '"Helvetica Neue", sans-serif'
+      system = '-apple-system, BlinkMacSystemFont, "Avenir Next", "Avenir", ' +
+               '"Segoe UI", "Lucida Grande", "Helvetica Neue", "Helvetica", ' +
+               '"Fira Sans", "Roboto", "Noto", "Droid Sans", "Cantarell", ' +
+               '"Oxygen", "Ubuntu", "Franklin Gothic Medium", ' +
+               '"Century Gothic", "Liberation Sans", sans-serif'
 
       expect(".helvetica").to have_value(helvetica)
       expect(".lucida-grande").to have_value(lucida_grande)